requirements.txt
flask==2.2.2
psycopg
psycopg-binary
python-dotenv
Flask-APScheduler